fastapi/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Shrivatsa"]

[tool.poetry.dependencies]
opentelemetry-api = "*"
opentelemetry-sdk = "*"
opentelemetry-exporter-zipkin = "*"
opentelemetry-exporter-prometheus = "*"
opentelemetry-instrumentation-fastapi = "*"
opentelemetry-instrumentation-aiohttp-client = "*"
httpx = "0.16"
python = "^3.7"
fastapi = "^0.62.0"
uvicorn = "^0.12.3"
python-dotenv = "^0.15.0"
python-consul2 = "0.1.5"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12", "setuptools>=46.0.0"]
build-backend = "poetry.masonry.api"
